Add files via upload
Not the final version. It can, for the most part, stand against a human player, but sometimes there are a few bugs, such as when the bot can win in it's next move but decides to stop an almost complete three in a row by the player. I am also still looking for a better way to improve other functions in general.