[PATCH] Difficulty setting, announce winner, cleaner look
Added a difficulty setting. It's still a WIP, but as of now it just sets a value that is used to lower the default depth of search in computerMove.
Added a function to announce the winner at the end of the game.
Changed pretty much every int to a short.
Changed the look of the board so it looks cleaner and is better to use / more appealing in general.